Fixed threading issue.
Reverted back to the old version, as it ended up being an issue with the q, not our code.
Removed David's branch attempt to fix it using thread persistence, as it was no longer necesary.
Removed the old debugging data, and added some real data from runs on the q.
Updated the output and plotting so that it would only show graphs for execution time, as requested by the updated spec.
Began working on the writeup.